Brief documentation added and init dropdown values
A brief documentation was added in the ConsultViewController for each action I
made.

The Campus, Building and Circuit dropdowns were initialize with the first item,
it means that the the campus dropdown has the first item selected and the
displayed text also has the name of the first item. The same for the first
building of the first campus and the first circuit of the first building.